requirements.txt
flask>=3.1.2
pillow>=10.1.0
pytest>=8.0.0
